refactor: use sub-commands, `Execute` -> `Prepare` in `Manager`
This is mostly just refactoring to make it easier to use multiple
subcommands with cobra, such that we can have `concierge prepare`,
and `concierge restore`.

This also renames `manager.Execute` -> `manager.Prepare`
