+++ internal/config/config_format.go
@@ -8,6 +8,7 @@ type Config struct {
 
 	// The following are added at runtime according to CLI flags
 	Overrides ConfigOverrides `mapstructure:"overrides"`
+	Verbose   bool            `mapstructure:"verbose"`
 }
 
 // jujuConfig represents the configuration for juju, including the desired version,
